# CLN: Remove PY2 compat code