- [ ] closes #xxxx
- [ ] tests added / passed
- [x] passes `black pandas`
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

---

Benchmarks:

#### Master:

```
In [1]: import pandas as pd                                                                                   

In [2]: s = pd.Series(list("ABCA"))                                                                           

In [3]: %timeit pd.get_dummies(s)                                                                             
482 µs ± 752 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

---

#### PR:

```
In [1]: import pandas as pd                                                                                   

In [2]: s = pd.Series(list("ABCA"))                                                                           

In [3]: %timeit pd.get_dummies(s)                                                                             
479 µs ± 2.43 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```